refactor: reorganize IOBuf implementation and remove event methods
