+++ sese/CMakeLists.txt
@@ -155,7 +155,7 @@ endif()
 # Windows platform configuration
 # ######################################################################################################################
 if(${CMAKE_SYSTEM_NAME} MATCHES "Windows")
-    file(GLOB_RECURSE NATIVE_WIN_SRC "internal/win/*.cpp" "native/win/*.cpp" "native/win/*.h")
+    file(GLOB_RECURSE NATIVE_WIN_SRC "internal/win/*.cpp" "internal/win/*.h")
 
     target_sources(Core PRIVATE ${NATIVE_WIN_SRC})
     target_compile_definitions(Core PRIVATE -D_WIN32_WINNT=0x0601)
@@ -182,8 +182,7 @@ if(${CMAKE_SYSTEM_NAME} MATCHES "Linux")
     target_link_libraries(Core PRIVATE ${CMAKE_THREAD_LIBS_INIT})
     target_link_libraries(Core PRIVATE ${CMAKE_DL_LIBS})
 
-    file(GLOB_RECURSE NATIVE_LINUX_SRC "internal/linux/*.cpp" "internal/linux/*.h" "native/linux/*.cpp"
-         "native/linux/*.h"
+    file(GLOB_RECURSE NATIVE_LINUX_SRC "internal/linux/*.cpp" "internal/linux/*.h"
     )
     target_sources(Core PRIVATE ${NATIVE_LINUX_SRC})
 
@@ -197,8 +196,7 @@ endif()
 if(${CMAKE_SYSTEM_NAME} MATCHES "Darwin")
     target_link_libraries(Core PRIVATE ${COREFOUNDATION_FRAMEWORK} ${IOKIT_FRAMEWORK} ${CORESERVICES_FRAMEWORK})
 
-    file(GLOB_RECURSE NATIVE_DARWIN_SRC "internal/darwin/*.cpp" "internal/darwin/*.h" "native/darwin/*.cpp"
-         "native/darwin/*.h"
+    file(GLOB_RECURSE NATIVE_DARWIN_SRC "internal/darwin/*.cpp" "internal/darwin/*.h"
     )
     target_sources(Core PRIVATE ${NATIVE_DARWIN_SRC})
 endif()
