+++ sese/example/CMakeLists.txt
@@ -7,10 +7,6 @@ include(${PROJECT_SOURCE_DIR}/cmake/Manifest.cmake)
 add_definitions("-DPROJECT_PATH=\"${PROJECT_SOURCE_DIR}\"")
 add_definitions("-DPROJECT_BINARY_PATH=\"${CMAKE_CURRENT_BINARY_DIR}\"")
 
-add_executable(EchoServer EchoServer.cpp)
-target_link_libraries(EchoServer PRIVATE Core)
-target_manifest(EchoServer manifest.json)
-
 add_executable(HttpServer HttpServer.cpp)
 target_link_libraries(HttpServer PRIVATE Core)
 target_manifest(HttpServer manifest.json)
