--- sese/internal/linux/net/event/EpollEventConvert.cpp
@@ -1,45 +1,0 @@
-// Copyright 2024 libsese
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#include "sese/internal/linux/net/event/EpollEventConvert.h"
-
-#include <sys/epoll.h>
-
-unsigned int sese::event::EpollEventConvert::fromNativeEvent(int event) {
-    short result = 0;
-    // if (event & EPOLLIN) {
-    //     result |= EVENT_READ;
-    // }
-    if (event & EPOLLOUT) {
-        result |= EVENT_WRITE;
-    }
-    if (event & EPOLLERR) {
-        result |= EVENT_ERROR;
-    }
-    return result;
-}
-
-int sese::event::EpollEventConvert::toNativeEvent(unsigned int event) {
-    int result = 0;
-    if (event & EVENT_READ) {
-        result |= EPOLLIN;
-    }
-    if (event & EVENT_WRITE) {
-        result |= EPOLLOUT;
-    }
-    if (event & EVENT_ERROR) {
-        result |= EPOLLERR;
-    }
-    return result;
-}
