--- sese/internal/win/net/event/WSAEventConvert.cpp
@@ -1,39 +1,0 @@
-// Copyright 2024 libsese
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#include "sese/internal/win/net/event/WSAEventConvert.h"
-
-#include <winsock2.h>
-
-unsigned int sese::event::WSAEventConvert::fromNativeEvent(int event) {
-    unsigned result = 0;
-    if (event & FD_READ) {
-        result |= EVENT_READ;
-    }
-    if (event & FD_WRITE) {
-        result |= EVENT_WRITE;
-    }
-    return result;
-}
-
-int sese::event::WSAEventConvert::toNativeEvent(unsigned int event) {
-    int result = 0;
-    if (event & EVENT_READ) {
-        result |= FD_READ;
-    }
-    if (event & EVENT_WRITE) {
-        result |= FD_WRITE;
-    }
-    return result;
-}
