--- sese/internal/win/net/event/WSAEventConvert.h
@@ -1,33 +1,0 @@
-// Copyright 2024 libsese
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-/**
- * @file WSAEventConvert.h
- * @brief WSAEventSelect event convert
- * @author kaoru
- */
-
-#pragma once
-
-#include "sese/net/event/BaseEventConvert.h"
-
-namespace sese::event {
-/// WSAEventSelect event convert
-class WSAEventConvert : public BaseEventConvert {
-public:
-    unsigned int fromNativeEvent(int event) override;
-
-    int toNativeEvent(unsigned int event) override;
-};
-} // namespace sese::event
