+++ sese/test/TestIOBuf.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <sese/service/iocp/IOBuf.h>
+#include <sese/util/IOBuf.h>
 
 #include <gtest/gtest.h>
