+++ sese/test/TestWebsocketService.cpp
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <sese/service/BalanceLoader.h>
-#include <sese/system/Process.h>
 #include <sese/record/Marco.h>
 #include <sese/net/ws/WebsocketAuthenticator.h>
